https://issues.apache.org/activemq/browse/AMQNET-172
Adds ConnectionInterrupted and ConnectionResumed events to the IConnection interface.  Clients will need to implement these and notify a client when a fault tolerant connection with a broker is broken / restored.  If the client doesn't provide fault tolerant connections then it is not required to send any notifications on these events.

git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-dotnet@825261 13f79535-47bb-0310-9956-ffa450edef68