https://issues.apache.org/activemq/browse/AMQNET-179
Remove the try/catch in the TcpTransports oneway method and just throw the error.  Error on send is in all cases treated as non-recoverable so throwing here is the right thing to do, the wrapped transports each get a change to process the error and remove any resources.  Connection catches errors on send and ensures that only NMSExceptions are thrown from there.  When used in a Failover transport this allows the logic in the oneway method to work more effectively instead of relying on the exception listener callback to trip the reconnect logic.

git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-dotnet@826034 13f79535-47bb-0310-9956-ffa450edef68