http://issues.apache.org/activemq/browse/AMQNET-220
Attempting to fix redelivery issues.  Found some things that the client was not doing correctly when sending various messages.  This submission adds the option the set a session to send acks async or sync, the session will also look at the type of session ack mode and for Transacted sessions always send the acks sync.  Added the DispatchAsync option to Connection to allow the connection to specify if consumers should configure themselves with the broker as asyncDispatch or not, default is true,  Also cleaned up some logging that duplicates what can be achieved by using the transport.UseLogging=true option.

Still have issue with message getting tagged on the broker as redelived when the actually aren't

git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-dotnet@891874 13f79535-47bb-0310-9956-ffa450edef68