http://issues.apache.org/activemq/browse/AMQNET-227
Improved ConnectionFactory to allow options to be set programmatically as well as through the URI.  The URI is now parsed on creation of the ConnectionFactory and options are set on parse, options set after the URI is parsed override the URI values.  IdGenerator is added to create Connection and Client Ids that match those in the Java client for better interoperability.

git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-dotnet@893218 13f79535-47bb-0310-9956-ffa450edef68