https://issues.apache.org/activemq/browse/AMQNET-261
Updates to the Uri Parsing and Composite Uri handling.  Allows most of the various forms of URIs that the AMQ docs call out.  

Added several new unit tests for URISupport and Tests in NMS.ActiveMQ to ensure that the Connection params are getting set even when a composite Uri is used for Failover transport.  

git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-dotnet@964638 13f79535-47bb-0310-9956-ffa450edef68