https://issues.apache.org/activemq/browse/AMQNET-284
Initial changes for Stomp v1.1 adds the proper headers to the CONNECT frame to activate 1.1 on brokers that support it, and reads off the headers from CONNECTED when 1.1 is indicated.  Adds the Read check to the current Inactivity monitor.  Handles multiple header values by keeping only the first one read and discarding the rest.

git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-dotnet@1031862 13f79535-47bb-0310-9956-ffa450edef68