https://issues.apache.org/jira/browse/AMQNET-290
Refactor how exceptions are dispatched in the Connection class to exception listeners, now done in a separate thread to prevent lock contention as the Transport is Disposed.  The Session is updated to not enlist a rollback operation if in a .NET style TX when the session is closed inside the TransactionScope.  Message Consumer adds a field to track failure of its parent connection so that calls to receive break when done after the Connection has received an OnException call.  When an error occurs on .NET transactions during Commit or Rollback its treated as an error that causes the parent connection to be closed as the Connection needs to be recreated so that an attempt to recover the TX can be done on startup.  

Adds a small fix to attempt to get the InactivityMonitor to stop leaking Timer threads, seems that the Timer class is sensitive and if Disoposed to quickly after creation / start it won't be shutdown.

git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-dotnet@1059208 13f79535-47bb-0310-9956-ffa450edef68